Elimination of some numerical problems.

+++ aabb.cpp
@@ -1,11 +1,11 @@
 #include "aabb.h"
 
-double AABB::getArea( void ) const
+float AABB::getArea( void ) const
 {
-    double l = ( max_.x - min_.x );
-    double m = ( max_.y - min_.y );
-    double n = ( max_.z - min_.z );
-    return 2.0 * ( l * ( m + n ) + m * n );
+    float l = ( max_.x - min_.x );
+    float m = ( max_.y - min_.y );
+    float n = ( max_.z - min_.z );
+    return 2.0f * ( l * ( m + n ) + m * n );
 }
 
 bool AABB::intersect( const Ray &ray ) const
@@ -16,18 +16,18 @@ bool AABB::intersect( const Ray &ray ) const
     // An Efficient and Robust Ray–Box Intersection Algorithm,
     //     Amy Williams, Steve Barrus, R. Keith Morley and Peter Shirley
 
-    double tmin;
-    double tmax;
-    double tymin;
-    double tymax;
-    double tzmin;
-    double tzmax;
-    double divx;
-    double divy;
-    double divz;
+    float tmin;
+    float tmax;
+    float tymin;
+    float tymax;
+    float tzmin;
+    float tzmax;
+    float divx;
+    float divy;
+    float divz;
 
-    divx = 1.0 / ray.direction_.x;
-    if ( divx >= 0.0 )
+    divx = 1.0f / ray.direction_.x;
+    if ( divx >= 0.0f )
     {
         tmin = ( min_.x - ray.origin_.x ) * divx;
         tmax = ( max_.x - ray.origin_.x ) * divx;
@@ -38,8 +38,8 @@ bool AABB::intersect( const Ray &ray ) const
         tmax = ( min_.x - ray.origin_.x ) * divx;
     }
 
-    divy = 1.0 / ray.direction_.y;
-    if ( divy >= 0.0 )
+    divy = 1.0f / ray.direction_.y;
+    if ( divy >= 0.0f )
     {
         tymin = ( min_.y - ray.origin_.y ) * divy;
         tymax = ( max_.y - ray.origin_.y ) * divy;
@@ -59,8 +59,8 @@ bool AABB::intersect( const Ray &ray ) const
     if ( tymax < tmax )
         tmax = tymax;
 
-    divz = 1.0 / ray.direction_.z;
-    if ( divz >= 0.0 )
+    divz = 1.0f / ray.direction_.z;
+    if ( divz >= 0.0f )
     {
         tzmin = ( min_.z - ray.origin_.z ) * divz;
         tzmax = ( max_.z - ray.origin_.z ) * divz;
@@ -85,5 +85,5 @@ bool AABB::intersect( const Ray &ray ) const
     //intersection_record.normal_ = normal_;
     //intersection_record.material_ = material_;
 
-    return true;//tmax > 0.00001;
+    return true;//tmax > 0.00001f;
 }
