+++ bsdf.h
@@ -10,14 +10,14 @@ public:
 
     using BSDFUniquePtr = std::unique_ptr< BSDF >;
 
-    BSDF( RNG< std::uniform_real_distribution, double, std::mt19937 > &rng );
+    BSDF( RNG< std::uniform_real_distribution, float, std::mt19937 > &rng );
 
-    glm::dvec3 fr( const glm::dvec3 &w_i,
-                   const std::vector< glm::dvec3 > &w_r ) const;
+    glm::vec3 fr( const glm::vec3 &w_i,
+                   const std::vector< glm::vec3 > &w_r ) const;
 
-    void getNewDirection( const glm::dvec3 &w_i, std::vector< glm::dvec3 > &w_r );
+    void getNewDirection( const glm::vec3 &w_i, std::vector< glm::vec3 > &w_r );
 
-    RNG< std::uniform_real_distribution, double, std::mt19937 > rng_;
+    RNG< std::uniform_real_distribution, float, std::mt19937 > rng_;
 
     std::vector< BxDF::BxDFUniquePtr > bxdf_layers_;
