+++ lua_bind.h
@@ -28,13 +28,13 @@ public:
                          Sampler **sampler,
                          Scene *scene,
                          Buffer **rendering_buffer,
-                         glm::dvec3 &background_color,
+                         glm::vec3 &background_color,
                          std::string &path_termination,
                          //std::size_t &max_path_depth,
                          std::size_t &path_length,
                          std::string &output_filename,
                          Scene::AccelerationStructure &scene_acceleration_data_structure,
-                         RNG< std::uniform_real_distribution, double, std::mt19937 > &rng );
+                         RNG< std::uniform_real_distribution, float, std::mt19937 > &rng );
 
 private:
 
@@ -43,11 +43,11 @@ private:
     void getCamera( Camera **camera );
 
     void getSampler( Sampler **sampler,
-                     RNG< std::uniform_real_distribution, double, std::mt19937 > &rng );
+                     RNG< std::uniform_real_distribution, float, std::mt19937 > &rng );
 
     void getBuffer( Buffer **rendering_buffer );
 
-    void getGlobals( glm::dvec3 &background_color,
+    void getGlobals( glm::vec3 &background_color,
                      std::string &path_termination,
                      //std::size_t &max_path_depth,
                      std::size_t &path_length,
@@ -55,39 +55,39 @@ private:
                      Scene::AccelerationStructure &scene_acceleration_data_structure );
 
     void getMaterial( Scene *scene,
-                      RNG< std::uniform_real_distribution, double, std::mt19937 > &rng );
+                      RNG< std::uniform_real_distribution, float, std::mt19937 > &rng );
 
     void getTriangle( Scene *scene,
-                      RNG< std::uniform_real_distribution, double, std::mt19937 > &rng );
+                      RNG< std::uniform_real_distribution, float, std::mt19937 > &rng );
 
     void getSphere( Scene *scene,
-                    RNG< std::uniform_real_distribution, double, std::mt19937 > &rng );
+                    RNG< std::uniform_real_distribution, float, std::mt19937 > &rng );
 
     void getMesh( Scene *scene,
                   const Scene::MeshType mesh_type,
-                  RNG< std::uniform_real_distribution, double, std::mt19937 > &rng );
+                  RNG< std::uniform_real_distribution, float, std::mt19937 > &rng );
 
 
     void getElements( Camera **camera,
                       Sampler **sampler,
                       Scene *scene,
                       Buffer **rendering_buffer,
-                      glm::dvec3 &background_color,
+                      glm::vec3 &background_color,
                       std::string &path_termination,
                       //max_path_depth,
                       std::size_t &path_length,
                       std::string &output_filename,
                       Scene::AccelerationStructure &scene_acceleration_data_structure,
-                      RNG< std::uniform_real_distribution, double, std::mt19937 > &rng );
+                      RNG< std::uniform_real_distribution, float, std::mt19937 > &rng );
 
     std::string parseString( const std::string &s_input );
 
-    double parseScalar( const std::string &s );
+    float parseScalar( const std::string &s );
 
-    glm::dvec3 parseVec3( const std::string &s );
+    glm::vec3 parseVec3( const std::string &s );
 
     void parseVertices( const std::string &s,
-                        glm::dvec3 *v );
+                        glm::vec3 *v );
 
 };
