+++ material.cpp
@@ -4,7 +4,7 @@ Material::Material( void )
 { }
 
 Material::Material( BSDF::BSDFUniquePtr bsdf,
-                    const glm::dvec3 &emitted ) :
+                    const glm::vec3 &emitted ) :
         bsdf_( std::move( bsdf ) ),
         emitted_{ emitted }
 { }
