+++ onb.cpp
@@ -3,8 +3,8 @@
 ONB::ONB( void )
 { }
 
-void ONB::setFromUW( const glm::dvec3 &u,
-                     const glm::dvec3 &w )
+void ONB::setFromUW( const glm::vec3 &u,
+                     const glm::vec3 &w )
 {
     u_ = u;
     w_ = w;
@@ -13,14 +13,14 @@ void ONB::setFromUW( const glm::dvec3 &u,
     setBasisMatrix();
 }
 
-void ONB::setFromV( const glm::dvec3 &v )
+void ONB::setFromV( const glm::vec3 &v )
 {
     v_ = v;
 
     if( fabs( v_.x ) > fabs( v_.y ) )
-        w_ = glm::normalize( glm::dvec3{ v_.z, 0.0, -v_.x } );
+        w_ = glm::normalize( glm::vec3{ v_.z, 0.0f, -v_.x } );
     else
-        w_ =  glm::normalize( glm::dvec3{ 0.0, -v_.z, v_.y } );
+        w_ =  glm::normalize( glm::vec3{ 0.0f, -v_.z, v_.y } );
 
     u_ = glm::cross( v_, w_ );
